% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_functions.R
\name{add_parenthesis}
\alias{add_parenthesis}
\title{Add parentheses around elements in a vector}
\usage{
add_parenthesis(vector)
}
\arguments{
\item{vector}{A vector of elements}
}
\value{
A vector with each element enclosed in parentheses
}
\description{
Add parentheses around elements in a vector
}
