% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_functions.R
\name{all_forest_template}
\alias{all_forest_template}
\title{Generate a forest plot for all data}
\usage{
all_forest_template(df_all, label_pos = 1.15)
}
\arguments{
\item{df_all}{Data frame containing all data}

\item{label_pos}{Position of the labels}
}
\value{
A ggplot object for the forest plot
}
\description{
Generate a forest plot for all data
}
