% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_robust.R
\name{count_and_robust}
\alias{count_and_robust}
\title{Count Unique Observations and Perform Robust Meta-Analysis}
\usage{
count_and_robust(data)
}
\arguments{
\item{data}{The dataset or subset for which to count unique observations and perform meta-analysis.}
}
\value{
A tibble with the count of distinct observations and meta-analysis results.
The tibble includes columns for unique observation count, beta coefficient, standard error (se),
and p-value for the meta-analysis.
}
\description{
Combines the functionality of counting unique studies and
performing a robust meta-analysis on a dataset. It is designed to streamline the process
of obtaining both counts and meta-analysis results for subsets of a dataset.
The function utilizes the \code{study_count} function to count the number of unique observations,
and the \code{map_robust} function to perform a robust meta-analysis on the given dataset.
}
