% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_calc.R
\name{d_calc}
\alias{d_calc}
\title{Calculate Cohen's D or Glass's ∆}
\usage{
d_calc(stat_type, stat, sample_sd, n_t, n_c)
}
\arguments{
\item{stat_type}{Category of statistical result reported or derived from the paper.
Possible values include "d_i_d", "d_i_m", "d_i_p", "d", "unspecified null",
"reg_coef", "t_test", "f_test", "odds_ratio", "log_odds_ratio".}

\item{stat}{Unstandardized effect size.}

\item{sample_sd}{Standard deviation of the relevant sample (preferably of the control group).}

\item{n_t}{Treatment group sample size.}

\item{n_c}{Control group sample size.}
}
\value{
Cohen's D or Glass's ∆ (Delta) value.
}
\description{
This function converts raw effect sizes into Cohen's D or Glass's ∆
based on the measure of sample variance for standardization.
}
\details{
All calculations derived from Cooper, Hedges, and Valentine (2009), except
for difference in proportions, which, to the best of our knowledge, Don Green
came up with while we were working on \emph{Prejudice Reduction: Progress and Challenges}.
We elaborate more on this estimator in the paper's appendix.
}
