% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip_calc.R
\name{difference_in_proportion_calc}
\alias{difference_in_proportion_calc}
\title{Calculate Glass's Delta based on Difference in Proportions}
\usage{
difference_in_proportion_calc(p1, p2)
}
\arguments{
\item{p1}{Proportion in the treatment group.}

\item{p2}{Proportion in the control group.}
}
\value{
Glass's $\Delta$ value.
}
\description{
This function estimates Glass's $\Delta$ based on two proportions, calculating
the difference between the proportions and using the square root of the control
group's product of proportion and complement of the proportion.
}
