% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bib.R
\name{manage_references}
\alias{manage_references}
\title{Generate references from a dataset and manage BibTeX file}
\usage{
manage_references(dat_with_dois, bib_file = "./refs.bib")
}
\arguments{
\item{dat_with_dois}{The dataset containing DOIs (columns: title, author, year, doi, url)}

\item{bib_file}{The file path to the BibTeX file}
}
\value{
A list of functions to manage BibTeX entries
}
\description{
This function processes a dataset containing DOIs and manages a BibTeX file.
It extracts DOIs from the dataset and writes corresponding BibTeX entries into a file.
Additionally, it provides a function to add individual DOIs to the same BibTeX file.
}
