% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-meta-function.R
\name{meta_analyze}
\alias{meta_analyze}
\title{meta_analyze}
\usage{
meta_analyze(
  ...,
  treatment_size_eqlgrtr_than = 0,
  treatment_size_less_than = Inf,
  drop_clusters = FALSE,
  keep_data = FALSE,
  add_lai_data = FALSE,
  add_many_labs = FALSE,
  lab_only = FALSE,
  online_only = FALSE,
  field_only = FALSE,
  output_file = FALSE
)
}
\arguments{
\item{...}{Group variables}

\item{treatment_size_eqlgrtr_than}{Minimum treatment size}

\item{treatment_size_less_than}{Maximum treatment size}

\item{drop_clusters}{Drop clusters flag}

\item{keep_data}{Keep data flag}

\item{add_lai_data}{Add lai data flag}

\item{add_many_labs}{Add many labs flag}

\item{lab_only}{Lab only flag}

\item{online_only}{Online only flag}

\item{field_only}{Field only flag}

\item{output_file}{Output file flag}
}
\value{
Output data
}
\description{
Main meta analysis function for paper
}
\details{
This function serves three main purposes:
\enumerate{
\item Prepare the data into subgroups for \code{meta_engine}
\item Call \code{meta_engine} function where the main analysis is produced
\item Export the output into a .csv file with consistent naming conventions
}
}
