% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-meta-function.R
\name{meta_engine}
\alias{meta_engine}
\title{meta_engine}
\usage{
meta_engine(df, group_vars)
}
\arguments{
\item{df}{The dataset}

\item{group_vars}{Variables for grouping}
}
\value{
Data after meta-analysis
}
\description{
Internal function for \code{meta_analyze()} (defined below)
}
\details{
Here the data is nested into mini-datasets based on \code{group_vars}.
Data is extracted from mini-datasets and then meta-analysis is run.
Statistics are then extracted from the meta-analysis object.
}
