% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_to_d.R
\name{odds_ratio_to_d}
\alias{odds_ratio_to_d}
\title{Convert Odds Ratio to Cohen's D}
\usage{
odds_ratio_to_d(p1, p2)
}
\arguments{
\item{p1}{Numerator for the odds ratio.}

\item{p2}{Denominator for the odds ratio.}
}
\value{
Cohen's D value.
}
\description{
This function is utilized in the appendix to contrast the conventional method
of calculating Cohen's D from a reported odds ratio against our preferred method,
using the Difference in Proportions estimator discussed elsewhere.
}
