% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_functions.R
\name{prep_meta_data}
\alias{prep_meta_data}
\title{Prepare meta data for plotting}
\usage{
prep_meta_data(subfield_df, overall_df = overall, subfield_var)
}
\arguments{
\item{subfield_df}{Data frame for subfield}

\item{overall_df}{Overall data frame}

\item{subfield_var}{Variable representing subfield}
}
\value{
Prepared data for meta analysis plotting
}
\description{
Prepare meta data for plotting
}
