% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_commit_push.R
\name{render_commit_push}
\alias{render_commit_push}
\title{Render, commit, and push changes for an R Markdown document}
\usage{
render_commit_push(commit_message = "update manuscript", skip_push = TRUE)
}
\arguments{
\item{commit_message}{A commit message}

\item{skip_push}{Logical indicating whether to skip the Git push step. Default is TRUE.}
}
\value{
None
}
\description{
This function renders (knits) an R Markdown document, performs a Git
commit with a provided commit message, and optionally pushes the changes to the remote repository.
}
\examples{
\dontrun{
render_commit_push("Update manuscript")
}
}
