% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_functions.R
\name{save_forest}
\alias{save_forest}
\title{Save a forest plot as an image}
\usage{
save_forest(name, width = 13, height = 8)
}
\arguments{
\item{name}{Name of the file to save}

\item{width}{Width of the image}

\item{height}{Height of the image}
}
\value{
None
}
\description{
Save a forest plot as an image
}
