% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_lm.R
\name{sum_lm}
\alias{sum_lm}
\title{Print Core Results from summary(lm()) in a Neat Table}
\usage{
sum_lm(dataset, y = NULL, x = NULL, coefs_only = TRUE, dgts = 5)
}
\arguments{
\item{dataset}{The dataset to be used for regression analysis.}

\item{y}{The response variable in the linear model.}

\item{x}{The predictor variable in the linear model.}

\item{coefs_only}{Logical, indicating whether to display only the coefficients table (default: TRUE).}

\item{dgts}{Number of digits for rounding coefficients (default: 5).}
}
\value{
If coefs_only is TRUE, returns a rounded coefficients table; otherwise, returns the full summary.
}
\description{
This function translates the output of summary(lm()) into a format that can be used with piping,
such as \code{dat |> filter(some_var) |> sum_lm()} or \code{dat |> split(~some_var) |> map(sum_lm)},
presenting the results in an easily transcribable table.
}
\note{
\code{dat} is the default dataset name, but you can put in whatever
}
\examples{
 \dontrun{
PaluckMetaSOP::contact_data |> sum_lm()
library(dplyr); library(purrr)
PaluckMetaSOP::sv_data |> split(~study_design) |> map(sum_lm)
}

}
