% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tab.R
\name{sum_tab}
\alias{sum_tab}
\title{Create a Frequency Table for a Variable in a Data Frame}
\usage{
sum_tab(data, var_name)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{var_name}{The name of the variable/column to generate the frequency table.}
}
\value{
A table showing the frequency of each unique value in the specified variable.
#'
}
\description{
\code{sum_tab} takes a built in function (table) and turns it into something
that can be integrated into a sequence of pipes. As with \code{sum_lm},
this is useful for when you want to see many tables at once,
or put a few tables into a bigger table.
Writing this was surprisingly complicated -- it uses enquo() to capture
the unquoted variable name and !! to unquote it within the function.
}
\examples{
# simple example: entire dataset
PaluckMetaSOP::sv_data |> sum_tab(behavior_type)
}
