% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_d_calc.R
\name{var_d_calc}
\alias{var_d_calc}
\title{Calculate Variance of Cohen's D or Glass's Delta}
\usage{
var_d_calc(d, n_t, n_c)
}
\arguments{
\item{d}{Standardized effect size calculated using \code{d_calc.R}.}

\item{n_t}{Treatment sample size.}

\item{n_c}{Control group sample size.}
}
\value{
Variance of Cohen's D or Glass's Delta
}
\description{
This function computes the variance of Cohen's D or Glass's Delta based on the provided effect size
and sample sizes, following the equation from Cooper, Hedges, and Valentine (2009).
}
