% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_robust.R
\name{map_robust}
\alias{map_robust}
\title{Apply metafor::robust in a pipe sequence or to subsets using purrr's split/map}
\usage{
map_robust(x)
}
\arguments{
\item{x}{The dataset or subset to perform meta-analysis on.}
}
\value{
A tibble/data frame with meta-analysis results.
}
\description{
Performs robust meta-analysis on either data subsets or the whole dataset..
Its intended usage is in in context of the  split/map & pipes paradigm:
\code{dat |> split(~some_var) |> map(map_robust)}.
}
\details{
\code{dat |> map_robust()} will return the meta-analytic estimate for the whole dataset.

NOTE: If there is only one study in a cluster, this function will
just reproduce that study's meta-analytic estimate verbatim. This is good for
some circumstances but not others. Adapt to your needs.
}
