% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_robust.R
\name{map_robust}
\alias{map_robust}
\title{Apply metafor::robust in a pipe sequence or to subsets using purrr's split/map}
\usage{
map_robust(x)
}
\arguments{
\item{x}{The dataset or subset to perform meta-analysis on.}
}
\value{
A tibble/data frame with meta-analysis results.
}
\description{
This function simplifies the process of performing a robust meta-analysis on subsets of a dataset.
The typical usage involves piping a dataset into the split/map paradigm: \code{dat |> split(~some_var) |> map(map_robust)}.
Alternatively, \code{dat |> map_robust()} can also be used.
}
