% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaluckMetaSOP-package.R
\docType{package}
\name{PaluckMetaSOP-package}
\alias{PaluckMetaSOP-package}
\alias{PaluckMetaSOP}
\title{PaluckMetaSOP: Standard Operating Procedures for meta-analyses by the Paluck Lab}
\description{
This functions in this package implement the Paluck lab's approach to meta-analysis, e.g. for creating SMDs, variance, standard errors, for meta-analyzing many papers, and for writing papers with a meta-analytic compoment. For help, see `browseVignettes("PaluckMetaSOP")`.

This package provides functions for conducting meta-analyses,
with a focus on standard operating procedures developed by the Paluck Lab.

It currently includes six functions: \code{\link{d_calc}}, \code{\link{map_robust}}, \code{\link{study_count}},
\code{\link{sum_lm}}, \code{\link{sum_tab}}, and \code{\link{var_d_calc}}.

It also includes two datasets from the Paluck Lab's meta-analyses:
\code{\link{contact_data}} and  \code{\link{sv_data}}. These datasets
are used in the vingettes and function documentation for as examples of
how to use the functions.

The main documentation for this package is in its vignettes. The first
provides an overview of the meta-analytic procedure and what these functions
are meant to help with. The second is about converting a study's results into
an estimate of standardized mean difference. The third is about doing meta-analysis.
The fourth is about how we write meta-analysis papers.
}
\author{
\strong{Maintainer}: Seth Green \email{setgree@gmail.com} (\href{https://orcid.org/0000-0003-3909-1969}{ORCID})

}
\keyword{internal}
\keyword{package}
