% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{PaluckMetaSOP-package}
\alias{PaluckMetaSOP-package}
\alias{PaluckMetaSOP}
\title{PaluckMetaSOP: Standard Operating Procedures for Meta-Analysis}
\description{
This package provides a collection of functions and datasets for conducting meta-analyses,
with a focus on standard operating procedures developed by the Paluck Lab.

The package includes functions for calculating effect sizes, performing robust meta-analysis,
managing references, and more. Additionally, it provides datasets related to sexual violence interventions.

For detailed documentation of individual functions and datasets, please see the reference
pages generated by pkgdown.
}
\seealso{
\code{\link{d_calc}}, \code{\link{count_and_robust}}, \code{\link{manage_references}},
\code{\link{map_robust}}, \code{\link{study_count}}, \code{\link{sum_lm}}, \code{\link{sum_tab}},
\code{\link{sv_data}}
}
\keyword{package}
