% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaluckMetaSOP-package.R
\docType{package}
\name{PaluckMetaSOP-package}
\alias{PaluckMetaSOP-package}
\alias{PaluckMetaSOP}
\title{PaluckMetaSOP: Standard Operating Procedures for meta-analyses by the Paluck Lab}
\description{
This functions in this package implement the Paluck lab's approach to meta-analysis, e.g. for creating SMDs, variance, standard errors, for meta-analyzing many papers, and for writing papers with a meta-analytic compoment. For help, see `browseVignettes("PaluckMetaSOP")`.

This package provides a collection of functions and datasets for conducting meta-analyses,
with a focus on standard operating procedures developed by the Paluck Lab.

The package includes functions for calculating effect sizes, performing robust meta-analysis,
managing references, and more. Additionally, it provides datasets related to sexual violence interventions.

For detailed documentation of individual functions and datasets, please see the reference
pages generated by pkgdown.
}
\seealso{
\code{\link{d_calc}}, \code{\link{manage_references}},
\code{\link{map_robust}}, \code{\link{study_count}}, \code{\link{sum_lm}}, \code{\link{sum_tab}},
\code{\link{sv_data}}
}
\author{
\strong{Maintainer}: Seth Green \email{setgree@gmail.com} (\href{https://orcid.org/0000-0003-3909-1969}{ORCID})

}
\keyword{internal}
\keyword{package}
