% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_count.R
\name{study_count}
\alias{study_count}
\title{Count the Number of Studies in a Dataset or Subset}
\usage{
study_count(dat)
}
\arguments{
\item{dat}{The dataset or subset.}
}
\value{
A tibble with the count of distinct studies.
}
\description{
This function simplifies the process of counting the number of studies in a dataset
or a subset by avoiding repetitive use of \code{summarise('N (studies)' = n_distinct(unique_study_id))}.
}
