% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_lm.R
\name{sum_lm}
\alias{sum_lm}
\title{Print Core Results from summary(lm()) in a Neat Table}
\usage{
sum_lm(dataset, y = NULL, x = NULL, coefs_only = TRUE, dgts = 5)
}
\arguments{
\item{dataset}{The dataset to be used for regression analysis.}

\item{y}{The response variable in the linear model.}

\item{x}{The predictor variable in the linear model.}

\item{coefs_only}{Logical, indicating whether to display only the coefficients table (default: TRUE).}

\item{dgts}{Number of digits for rounding coefficients (default: 5).}
}
\value{
If coefs_only is TRUE, returns a rounded coefficients table; otherwise, returns the full summary.
}
\description{
\code{sum_lm} takes the built in R functions \code{summary(lm())}
and lets you use them in a sequence of pipes.
In particular this is useful when you want to test for
the magnitude of a relationship between two variables in
many different subsets of data at once.
}
\note{
\code{dat} is the default dataset name, but you can put in whatever
}
\examples{
# example 1: entire dataset
PaluckMetaSOP::contact_data |> sum_lm()
# example 2: split and apply to many subsets
\dontrun{
library(purrr)
PaluckMetaSOP::sv_data |> split(~study_design) |> map(sum_lm)
}
}
