% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tab.R
\name{sum_tab}
\alias{sum_tab}
\title{Create a Frequency Table for a Variable in a Data Frame}
\usage{
sum_tab(data, var_name)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{var_name}{The name of the variable/column to generate the frequency table.}
}
\value{
A table showing the frequency of each unique value in the specified variable.
#'
}
\description{
This function generates a frequency table for a  variable in a dataset.
It's equivalent to \code{table(data$some_var)}, but reworked so you can use it
in a sequence of pipes, e.g. \code{data |> filter(var_one) |> sum_tab(var_two)}.
This is useful when you want to use the split-apply-combine paradigm.
Writing this was surprisingly complicated -- it uses enquo() to capture
the unquoted variable name and !! to unquote it within the function.
}
\examples{
 \dontrun{
# simple example
PaluckMetaSOP::sv_data |> sum_tab(behavior_type)
# example with split by some other factor
library(purrr)
PaluckMetaSOP::sv_data |> split(~study_design) |>
map(~ sum_tab(., behavior_type)) |> bind_rows(.id = "study_design")
}

}
