% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tab.R
\name{sum_tab}
\alias{sum_tab}
\title{Create a Frequency Table for a Variable in a Data Frame}
\usage{
sum_tab(data, var_name)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{var_name}{The name of the variable/column to generate the frequency table.}
}
\value{
A table showing the frequency of each unique value in the specified variable.
#'
}
\description{
This function generates a frequency table for a specific variable in a dataset.
}
\details{
It uses enquo() to capture the unquoted variable name and !! to unquote it
within the function.
}
